+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- Add `downcast` API for `dyn Error` types
+
 ## [0.1.1] - 2024-08-30
 
 ### Added